requirements.txt
alabaster==0.7.12
alembic==1.4.2
anytree==2.8.0
appdirs==1.4.3
asn1crypto==1.3.0
astroid==2.3.3
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
backports.shutil-which==3.5.2
bcrypt==3.1.7
Beaker==1.11.0
blinker==1.4
botocore==1.17.54
breezy==3.0.2
Brlapi==0.7.0
btrfsutil==1.2.0
CacheControl==0.12.6
cached-property==1.5.1
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
ConfigArgParse==1.1
configobj==5.0.6
contextlib2==0.6.0.post1
cryptography==2.9.2
decorator==4.4.2
distlib==0.3.0
distro==1.5.0
docker==4.2.2
docker-compose==1.27.3
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.16
dulwich==0.19.16
ecdsa==0.15
ECPy==1.2.1
ed25519==1.5
extras==1.0.0
fastimport==0.9.8
fixtures==3.0.0
Flask==1.1.2
Flask-BabelEx==0.9.4
Flask-Compress==1.5.0
Flask-Gravatar==0.5.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security-Too==3.3.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
future==0.18.2
greenlet==0.4.15
hidapi==0.9.0.post2
html5lib==1.0.1
idna==2.9
imagesize==1.2.0
importlib-metadata==1.5.2
ipython==7.13.0
ipython-genutils==0.1.0
isc==2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
lazy-object-proxy==1.4.3
ldap3==2.8.1
-e git+https://github.com/romanz/trezor-agent@61cfcef35ca80737992f9c34bea49b55069d3536#egg=ledger_agent&subdirectory=agents/ledger
ledgerblue==0.1.31
lensfun==0.3.95
libagent==0.13.1
lockfile==0.12.2
louis==3.13.0
lxml==4.5.0
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
meson==0.54.1
mnemonic==0.19
msgpack==1.0.0
ordered-set==3.1.1
packaging==20.3
paramiko==2.7.1
parso==0.6.2
passlib==1.7.2
patiencediff==0.1.0
pbr==5.5.0
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
ply==3.11
progress==1.5
prompt-toolkit==3.0.5
protobuf==3.11.3
psutil==5.7.2
psycopg2==2.8.6
ptyprocess==0.6.0
pwquality==1.4.2
pyasn1==0.4.8
pycairo==1.19.1
pycodestyle==2.5.0
pycparser==2.20
pycryptodomex==3.9.7
pyflakes==2.2.0
Pygments==2.6.1
PyGObject==3.36.0
pylint==2.4.4
PyMsgBox==1.0.7
PyNaCl==1.3.0
pynvim==0.4.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-daemon==2.2.4
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-mimeparse==1.6.0
python-u2flib-host==3.0.3
python-xlib==0.27
pytoml==0.1.21
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
rados==2.0.0
rbd==2.0.0
requests==2.23.0
retrying==1.3.3
rgw==2.0.0
rsa==4.6
s3transfer==0.3.3
semver==2.9.1
simplejson==3.17.0
six==1.14.0
snowballstemmer==2.0.0
speaklater==1.3
Sphinx==3.0.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.3.19
sqlparse==0.3.1
sshtunnel==0.1.5
team==1.0
testtools==2.4.0
texttable==1.6.2
toml==0.10.0
traitlets==4.3.3
Unidecode==1.1.1
urllib3==1.25.9
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.2.1
zipp==3.1.0
